Channels: Auto create stream associated
When requesting a new remote channel remotely, data could arrive before
local was created and cause program to error because it tried to write
into a closed (really not yet opened) channel.
Now on channel creation, a memory stream is associated to buffers data
before real stream is associated with.